azure-ai-security/requirements.txt
python-dotenv
azure-ai-textanalytics==5.3.0
azure-identity==1.5.0
azure-keyvault-secrets==4.2.0